refactor as shared library
